requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
attrs==21.2.0
cbpro==1.1.4
certifi==2020.12.5
chardet==4.0.0
dateparser==1.0.0
future==0.18.2
idna==2.10
multidict==5.1.0
pefile==2019.4.18
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pymongo==3.5.1
python-binance==1.0.14
python-dateutil==2.8.2
pytz==2021.1
pywin32-ctypes==0.2.0
regex==2021.8.28
requests==2.25.1
six==1.10.0
sortedcontainers==2.3.0
typing-extensions==3.10.0.2
tzdata==2021.1
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.2
websocket-client==0.40.0
websockets==10.0
yarl==1.6.3